separate anvil stuff from mod.rs in chunk/
