+++ pumpkin-world/src/chunk/format/anvil.rs
@@ -2,18 +2,101 @@ use fastnbt::LongArray;
 use indexmap::IndexMap;
 use pumpkin_util::math::ceil_log2;
 use pumpkin_util::math::vector2::Vector2;
-use std::collections::HashSet;
+use serde::{Deserialize, Serialize};
+use std::collections::{HashMap, HashSet};
+use thiserror::Error;
 
 use crate::block::block_registry::BLOCK_ID_TO_REGISTRY_ID;
+use crate::block::BlockState;
 use crate::chunk::{
-    ChunkData, ChunkNbt, ChunkSection, ChunkSectionBlockStates, ChunkStatus, PaletteEntry,
+    ChunkData, ChunkHeightmaps, ChunkParsingError, Subchunks, CHUNK_AREA, SUBCHUNK_VOLUME,
     WORLD_DATA_VERSION,
 };
+use crate::coordinates::{ChunkRelativeBlockCoordinates, Height};
 
 use super::{ChunkReader, ChunkReadingError, ChunkWriter, ChunkWritingError};
 
 #[derive(Clone, Default)]
 pub struct AnvilChunkFormat;
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+#[serde(rename_all = "PascalCase")]
+struct PaletteEntry {
+    // block name
+    name: String,
+    properties: Option<HashMap<String, String>>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct ChunkSection {
+    #[serde(rename = "Y")]
+    y: i8,
+    block_states: Option<ChunkSectionBlockStates>,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+struct ChunkSectionBlockStates {
+    //  #[serde(with = "LongArray")]
+    data: Option<LongArray>,
+    palette: Vec<PaletteEntry>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(rename_all = "PascalCase")]
+struct ChunkNbt {
+    data_version: i32,
+    #[serde(rename = "xPos")]
+    x_pos: i32,
+    // #[serde(rename = "yPos")]
+    //y_pos: i32,
+    #[serde(rename = "zPos")]
+    z_pos: i32,
+    status: ChunkStatus,
+    #[serde(rename = "sections")]
+    sections: Vec<ChunkSection>,
+    heightmaps: ChunkHeightmaps,
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
+enum ChunkStatus {
+    #[serde(rename = "minecraft:empty")]
+    Empty,
+    #[serde(rename = "minecraft:structure_starts")]
+    StructureStarts,
+    #[serde(rename = "minecraft:structure_references")]
+    StructureReferences,
+    #[serde(rename = "minecraft:biomes")]
+    Biomes,
+    #[serde(rename = "minecraft:noise")]
+    Noise,
+    #[serde(rename = "minecraft:surface")]
+    Surface,
+    #[serde(rename = "minecraft:carvers")]
+    Carvers,
+    #[serde(rename = "minecraft:features")]
+    Features,
+    #[serde(rename = "minecraft:initialize_light")]
+    InitLight,
+    #[serde(rename = "minecraft:light")]
+    Light,
+    #[serde(rename = "minecraft:spawn")]
+    Spawn,
+    #[serde(rename = "minecraft:full")]
+    Full,
+}
+
+// I can't use an tag because it will break ChunkNBT, but status need to have a big S, so "Status"
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(rename_all = "PascalCase")]
+pub struct ChunkStatusWrapper {
+    status: ChunkStatus,
+}
+
+#[derive(Error, Debug)]
+pub enum ChunkSerializingError {
+    #[error("Error serializing chunk: {0}")]
+    ErrorSerializingChunk(fastnbt::error::Error),
+}
 
 impl ChunkReader for AnvilChunkFormat {
     fn read_chunk(
@@ -21,7 +104,103 @@ impl ChunkReader for AnvilChunkFormat {
         chunk_bytes: Vec<u8>,
         at: &Vector2<i32>,
     ) -> Result<ChunkData, ChunkReadingError> {
-        ChunkData::from_bytes(&chunk_bytes, *at).map_err(ChunkReadingError::ParsingError)
+        if fastnbt::from_bytes::<ChunkStatusWrapper>(&chunk_bytes)
+            .map_err(|_| ChunkReadingError::InvalidHeader)?
+            .status
+            != ChunkStatus::Full
+        {
+            return Err(ChunkReadingError::ChunkNotExist);
+        }
+
+        let chunk_data = fastnbt::from_bytes::<ChunkNbt>(&chunk_bytes).map_err(|e| {
+            ChunkReadingError::ParsingError(ChunkParsingError::ErrorDeserializingChunk(
+                e.to_string(),
+            ))
+        })?;
+
+        if chunk_data.x_pos != at.x || chunk_data.z_pos != at.z {
+            log::error!(
+                "Expected chunk at {}:{}, but got {}:{}",
+                at.x,
+                at.z,
+                chunk_data.x_pos,
+                chunk_data.z_pos
+            );
+            // lets still continue
+        }
+
+        let mut subchunks = Subchunks::Single(0);
+        let mut block_index = 0; // which block we're currently at
+
+        for section in chunk_data.sections.into_iter() {
+            let block_states = match section.block_states {
+                Some(states) => states,
+                None => continue, // TODO @lukas0008 this should instead fill all blocks with the only element of the palette
+            };
+
+            let palette = block_states
+                .palette
+                .iter()
+                .map(|entry| match BlockState::new(&entry.name) {
+                    // Block not found, Often the case when World has an newer or older version then block registry
+                    None => BlockState::AIR,
+                    Some(state) => state,
+                })
+                .collect::<Vec<_>>();
+
+            let block_data = match block_states.data {
+                None => {
+                    // We skipped placing an empty subchunk.
+                    // We need to increase the y coordinate of the next subchunk being placed.
+                    block_index += SUBCHUNK_VOLUME;
+                    continue;
+                }
+                Some(d) => d,
+            };
+
+            // How many bits each block has in one of the palette u64s
+            let block_bit_size = if palette.len() < 16 {
+                4
+            } else {
+                ceil_log2(palette.len() as u32).max(4)
+            };
+            // How many blocks there are in one of the palettes u64s
+            let blocks_in_palette = 64 / block_bit_size;
+
+            let mask = (1 << block_bit_size) - 1;
+            'block_loop: for block in block_data.iter() {
+                for i in 0..blocks_in_palette {
+                    let index = (block >> (i * block_bit_size)) & mask;
+                    let block = &palette[index as usize];
+
+                    // TODO allow indexing blocks directly so we can just use block_index and save some time?
+                    // this is fine because we initialized the heightmap of `blocks`
+                    // from the cached value in the world file
+                    subchunks.set_block_no_heightmap_update(
+                        ChunkRelativeBlockCoordinates {
+                            z: ((block_index % CHUNK_AREA) / 16).into(),
+                            y: Height::from_absolute((block_index / CHUNK_AREA) as u16),
+                            x: (block_index % 16).into(),
+                        },
+                        block.get_id(),
+                    );
+
+                    block_index += 1;
+
+                    // if `SUBCHUNK_VOLUME `is not divisible by `blocks_in_palette` the block_data
+                    // can sometimes spill into other subchunks. We avoid that by aborting early
+                    if (block_index % SUBCHUNK_VOLUME) == 0 {
+                        break 'block_loop;
+                    }
+                }
+            }
+        }
+
+        Ok(ChunkData {
+            subchunks,
+            heightmap: chunk_data.heightmaps,
+            position: *at,
+        })
     }
 }
